refactpr(frontend): shared map zoom state
Refactor the base map so that map zoom state is handled in `lib/state/map-view`. Initial map state and style is configured in `app/config/map-view` and passed into the base map. `lib/data-map/BaseMap` is now responsible for updating the map when zoom parameters change.
