+++ frontend/src/app/Nav.tsx
@@ -19,7 +19,7 @@ import { NavLink as RouterNavLink, LinkProps as RouterLinkProps } from 'react-ro
 import { VIEW_SECTIONS } from 'app/config/views';
 import { useIsMobile } from './use-is-mobile';
 import { withProps } from 'lib/react/with-props';
-import { mapLatUrlState, mapLonUrlState, mapZoomUrlState } from 'app/state/map-view/map-url';
+import { mapLatUrlState, mapLonUrlState, mapZoomUrlState } from 'lib/state/map-view/map-url';
 import { globalStyleVariables } from 'app/theme';
 import { useRecoilValue } from 'recoil';
