+++ frontend/src/app/map/BaseMap.tsx
@@ -1,10 +1,11 @@
-import { MapViewState } from 'deck.gl/typed';
+import omit from 'lodash/omit';
 import { FC, ReactNode } from 'react';
 import { useRecoilState, useRecoilValue } from 'recoil';
 
-import { mapViewStateState, useSyncMapUrl } from 'app/state/map-view/map-view-state';
+import { mapViewConfig } from 'app/config/map-view';
+import { mapViewStateState, nonCoordsMapViewStateState } from 'lib/state/map-view/map-view-state';
 import { BaseMap } from 'lib/data-map/BaseMap';
-import { labelZoomState } from 'lib/sidebar/ui/LayerLabel';
+
 import { backgroundState, showLabelsState } from './layers/layers-state';
 import { useBasemapStyle } from './use-basemap-style';
 
@@ -12,22 +13,24 @@ export interface BaseMapProps {
   children?: ReactNode;
 }
 
+const INITIAL_VIEW_STATE = {
+  ...mapViewConfig.initialViewState,
+  ...mapViewConfig.viewLimits,
+};
+const INITIAL_NON_COORDS_STATE = omit(INITIAL_VIEW_STATE, ['latitude', 'longitude', 'zoom']);
+
 export const BaseMapContainer: FC<BaseMapProps> = ({ children }) => {
   const background = useRecoilValue(backgroundState);
   const showLabels = useRecoilValue(showLabelsState);
   const [viewState, setViewState] = useRecoilState(mapViewStateState);
-  const [labelZoom, setLabelZoom] = useRecoilState(labelZoomState);
+  const [nonCoordsViewState, setNonCoordsViewState] = useRecoilState(nonCoordsMapViewStateState);
   const { mapStyle } = useBasemapStyle(background, showLabels);
-  useSyncMapUrl();
-
-  function handleViewStateChange({ viewState }: { viewState: MapViewState }) {
-    setViewState(viewState);
-    setLabelZoom(viewState.zoom);
+  if (viewState.zoom < 0) {
+    setViewState(INITIAL_VIEW_STATE);
+    setNonCoordsViewState(INITIAL_NON_COORDS_STATE);
+    // wait for initial state to be set before showing a map.
+    return null;
   }
 
-  return (
-    <BaseMap onMove={handleViewStateChange} viewState={viewState} mapStyle={mapStyle}>
-      {children}
-    </BaseMap>
-  );
+  return <BaseMap mapStyle={mapStyle}>{children}</BaseMap>;
 };
